pyproject.toml
[project]
name = "venv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coverage~=7.6",
    "django~=4.2",
    "django-cors-headers~=4.4",
    "django-filter>=25.1",
    "djangorestframework~=3.15",
    "djangorestframework-simplejwt~=5.3",
    "drf-spectacular~=0.27",
    "pre-commit~=4.2",
    "ruff==0.12.7",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "ASYNC",
    "T10",
    "ISC",
    "ICN",
    "PIE",
    "BLE",
    "SIM",
    "PTH",
    "FIX",
    "N",
    "UP",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403"]